Major Refactor - Refactor All views
Refactor all views so that they have their own file if appropriate.
