+++ DogBreeds/DogBreeds/ContentView.swift
@@ -15,7 +15,6 @@ struct ContentView: View {
     }
     
     @State private var tabSelection = Tab.Home
-
     
     var body: some View {
         
@@ -35,130 +34,3 @@ struct ContentView: View {
 }
 
 
-struct HomeScreenView: View {
-    
-    @State private var dogImageURLString = ""
-    @State private var firstStartUp = true
-
-    let defaultDogURLString = "https://images.dog.ceo/breeds/segugio-italian/n02090722_002.jpg"
-    let heroMessage = "Can you recognize different dog breeds?"
-    let heroSubMessage = "Search for breeds and see pictures for familiarity. Or simply tap the button below for a random dog image!"
-    
-    
-    var body: some View {
-        ScrollView {
-            Image(.header)
-                .padding(.top, 30)
-            Spacer()
-            Image(.hero)
-                .resizable()
-                .scaledToFill()
-                .clipShape(.rect(cornerRadius: 10))
-                .padding(.horizontal, 5)
-                .padding(.vertical)
-            VStack{
-                Text(heroMessage)
-                    .font(.largeTitle)
-                    .padding(.bottom)
-                Text(heroSubMessage)
-                    .font(.headline)
-                    .padding(.horizontal)
-            }
-            // using Apple example of error handling for AsyncImage
-            VStack {
-                AsyncImage(url: URL(string: dogImageURLString )) { phase in
-                    if let image = phase.image {
-                        image
-                            .resizable()
-                            .frame(width: 300, height: 250)
-                            .cornerRadius(5)
-                        
-                    } else if phase.error != nil {
-                        Label("Could not load image", systemImage: "photo")
-                            .foregroundColor(Color.primary)
-                    } else {
-                        if !firstStartUp {
-                            ProgressView("Loading...")
-                                .foregroundColor(Color.primary)
-                        } else {
-                            EmptyView()
-                        }
-                        
-
-                    }
-                }// end AsyncImage
-                
-                Button {
-                    grabRandomDogImageURL()
-                    firstStartUp = false
-                } label: {
-                    
-                    Text("Next Dog 🐶")
-                        .foregroundStyle(Color.primary)
-                        .frame(maxWidth: 150)
-                }
-                .buttonStyle(.bordered)
-                .foregroundColor(Color.black)
-                .cornerRadius(5)
-                .padding(.vertical)
-            } // end VStack
-            
-
-            Spacer()
-        } // end ScrollView
-      //  .onAppear{ grabRandomDogImageURL() }
-        
-        
-    } // end View
-        
-        func grabRandomDogImageURL() {
-            
-            // get JSON URL
-            let url = URL(string: "https://dog.ceo/api/breeds/image/random")!
-            
-            // create URL request
-            let urlRequest = URLRequest(url: url)
-            
-            // create URL Session task
-            let task = URLSession.shared.dataTask(with: urlRequest) { data, reponse, error in
-                
-                if let data = data, let string = String(data: data, encoding: .utf8) {
-                    
-                    // convert string to Data format
-                    let DogData = Data(string.utf8)
-                    
-                    // create decoder and use it to decode based off of RandomDogImageItem
-                    let decoder = JSONDecoder()
-                    let dogImageURLItem = try! decoder.decode(RandomDogImageItem.self, from: DogData)
-                    
-                    //print(dogImageURLString.message)
-                    //dogImage = URL(string: dogImageURLString.message)
-                    dogImageURLString = dogImageURLItem.message
-                    
-                } else {
-                    print("unsuccessful decoding - data not valid")
-                }
-            }
-            task.resume()
-        } // end function
-        
-    }
-    
-    struct DogBreedsView: View {
-        var body: some View {
-            Text("Coming Soon!")
-        }
-    }
-    
-    struct RandomDogView: View {
-        var body: some View {
-            Text("Coming Soon!")
-        }
-    }
-    
-    
-    // must decode dog string first
-    struct RandomDogImageItem: Decodable {
-        let message: String
-        let status: String
-    }
