+++ DogBreeds/DogBreeds/DogBreedsView.swift
@@ -1,0 +1,18 @@
+//
+//  DogBreedsView.swift
+//  DogBreeds
+//
+//  Created by Joel Sereno on 10/21/23.
+//
+
+import SwiftUI
+
+struct DogBreedsView: View {
+    var body: some View {
+        Text("Coming Soon!")
+    }
+}
+
+#Preview {
+    DogBreedsView()
+}
