+++ DogBreeds/DogBreeds/DogData.swift
@@ -1,0 +1,14 @@
+//
+//  DogData.swift
+//  DogBreeds
+//
+//  Created by Joel Sereno on 10/21/23.
+//
+
+import Foundation
+
+// Needed to decode random dog image
+struct RandomDogImageItem: Decodable {
+    let message: String
+    let status: String
+}
